Request zram state via zramctl
On some configurations (like these based on 3.x kernel) there is no /sys/class/zram-control being created on zram init. It seems a good idea to do the check using zramctl tool, which still works fine. It should not decrease the compatibility of the script in any way as it already requires zramctl elsewhere.